+++ cmd/init.go
@@ -52,5 +52,6 @@ func PrintServerInit(servers []dao.Server) {
 	}
 
 	fmt.Println("\nFollowing servers were added to sake.yaml")
-	print.PrintTable(data.Rows, options, data.Headers)
+	err := print.PrintTable(data.Rows, options, data.Headers)
+	core.CheckIfError(err)
 }