+++ core/run/text.go
@@ -21,23 +21,43 @@ import (
 )
 
 func (run *Run) Text(dryRun bool) error {
+	task := run.Task
 	servers := run.Servers
 	prefixMaxLen := calcMaxPrefixLength(run.LocalClients)
 
+	var dataExit dao.TableOutput
+	var dataMutex = sync.RWMutex{}
+	dataExit.Headers = append(dataExit.Headers, "server")
+	// Append Command names if set
+	for _, subTask := range task.Tasks {
+		dataExit.Headers = append(dataExit.Headers, subTask.Name)
+	}
+	// Populate the rows (server name is first cell, then commands and cmd output is set to empty string)
+	for i, p := range servers {
+		dataExit.Rows = append(dataExit.Rows, dao.Row{Columns: []string{p.Name}})
+
+		for range task.Tasks {
+			dataExit.Rows[i].Columns = append(dataExit.Rows[i].Columns, "")
+		}
+	}
+
+	waitChan := make(chan struct{}, 100)
 	var wg sync.WaitGroup
 	for i := range servers {
 		wg.Add(1)
+		waitChan <- struct{}{}
 
 		if run.Task.Spec.Parallel {
 			go func(i int, wg *sync.WaitGroup) {
 				defer wg.Done()
 				// TODO: Handle errors when running tasks in parallel
-				_ = run.TextWork(i, prefixMaxLen, dryRun)
+				_ = run.TextWork(i, prefixMaxLen, dryRun, dataExit, &dataMutex)
+				<-waitChan
 			}(i, &wg)
 		} else {
 			err := func(i int, wg *sync.WaitGroup) error {
 				defer wg.Done()
-				err := run.TextWork(i, prefixMaxLen, dryRun)
+				err := run.TextWork(i, prefixMaxLen, dryRun, dataExit, &dataMutex)
 				return err
 			}(i, &wg)
 
@@ -69,7 +89,7 @@ func (run *Run) Text(dryRun bool) error {
 	return nil
 }
 
-func (run *Run) TextWork(rIndex int, prefixMaxLen int, dryRun bool) error {
+func (run *Run) TextWork(rIndex int, prefixMaxLen int, dryRun bool, dataExit dao.TableOutput, dataMutex *sync.RWMutex) error {
 	config := run.Config
 	task := run.Task
 	server := run.Servers[rIndex]
@@ -78,9 +98,11 @@ func (run *Run) TextWork(rIndex int, prefixMaxLen int, dryRun bool) error {
 	numTasks := len(task.Tasks)
 
 	var wg sync.WaitGroup
+	register := make(map[string]string)
+	var registers []string
 	for j, cmd := range task.Tasks {
 		var client Client
-		combinedEnvs := dao.MergeEnvs(cmd.Envs, server.Envs)
+		combinedEnvs := dao.MergeEnvs(cmd.Envs, server.Envs, registers)
 		if cmd.Local || server.Local {
 			client = run.LocalClients[server.Name]
 		} else {
@@ -105,7 +127,43 @@ func (run *Run) TextWork(rIndex int, prefixMaxLen int, dryRun bool) error {
 			tty:      cmd.TTY,
 		}
 
-		err := RunTextCmd(args, task.Theme.Text, prefix, task.Spec.Parallel, &wg)
+		buf, bufOut, bufErr, err := RunTextCmd(args, task.Theme.Text, prefix, task.Spec.Parallel, dataExit, task.Tasks[j].Register, dataMutex, &wg)
+
+		// Add exit code to dataExit
+		var errCode int
+		switch err := err.(type) {
+		case *ssh.ExitError:
+			errCode = err.ExitStatus()
+		case *exec.ExitError:
+			errCode = err.ExitCode()
+		}
+
+		dataExit.Rows[rIndex].Columns[j+1] = fmt.Sprint(errCode)
+
+		// variable
+		// stdout
+		// stderr
+		// rc
+		// failed
+		if task.Tasks[j].Register != "" {
+			register[task.Tasks[j].Register] = buf
+			register[task.Tasks[j].Register + "_stdout"] = bufOut
+			register[task.Tasks[j].Register + "_stderr"] = bufErr
+			register[task.Tasks[j].Register + "_rc"] = dataExit.Rows[rIndex].Columns[j+1]
+			if err != nil {
+				register[task.Tasks[j].Register + "_failed"] = "true"
+			} else {
+				register[task.Tasks[j].Register + "_failed"] = "false"
+			}
+			// TODO: Add skipped env variable
+
+			registers = []string{}
+			for  k, v := range register {
+				envStdout := fmt.Sprintf("%v=%v", k, v)
+				registers = append(registers, envStdout)
+			}
+		}
+
 		switch err.(type) {
 		case *template.ExecError:
 			return err
@@ -123,36 +181,52 @@ func (run *Run) TextWork(rIndex int, prefixMaxLen int, dryRun bool) error {
 	return nil
 }
 
-func RunTextCmd(t TaskContext, textStyle dao.Text, prefix string, parallel bool, wg *sync.WaitGroup) error {
+func RunTextCmd(t TaskContext, textStyle dao.Text, prefix string, parallel bool, dataExit dao.TableOutput, register string, dataMutex *sync.RWMutex, wg *sync.WaitGroup) (string, string, string, error) {
+	buf := new(bytes.Buffer)
+	bufOut := new(bytes.Buffer)
+	bufErr := new(bytes.Buffer)
+
 	if textStyle.Header != "" && !parallel {
 		err := printHeader(t.cIndex, t.numTasks, t.name, t.desc, textStyle)
 		if err != nil {
-			return err
+			return buf.String(), bufOut.String(), bufErr.String(), err
 		}
 	}
 
 	if t.dryRun {
 		printCmd(prefix, t.cmd)
-		return nil
+		return buf.String(), bufOut.String(), bufErr.String(), nil
 	}
 
 	if t.tty {
-		return ExecTTY(t.cmd, t.env)
+		return buf.String(), bufOut.String(), bufErr.String(), ExecTTY(t.cmd, t.env)
 	}
 
 	err := t.client.Run(t.env, t.workDir, t.shell, t.cmd)
 	if err != nil {
-		return err
+		return buf.String(), bufOut.String(), bufErr.String(), err
 	}
 
 	// Copy over commands STDOUT.
 	go func(client Client) {
 		defer wg.Done()
 		var err error
-		if prefix != "" {
-			_, err = io.Copy(os.Stdout, core.NewPrefixer(client.Stdout(), prefix))
+
+		if register == "" {
+			if prefix != "" {
+				_, err = io.Copy(os.Stdout, core.NewPrefixer(client.Stdout(), prefix))
+			} else {
+				_, err = io.Copy(os.Stdout, client.Stdout())
+			}
 		} else {
-			_, err = io.Copy(os.Stdout, client.Stdout())
+			mw := io.MultiWriter(buf, bufOut)
+			r := io.TeeReader(client.Stdout(), mw)
+			// TODO: Refactor to NewReader: https://pkg.go.dev/golang.org/x/text/transform?utm_source=godoc#NewReader
+			if prefix != "" {
+				_, err = io.Copy(os.Stdout, core.NewPrefixer(r, prefix))
+			} else {
+				_, err = io.Copy(os.Stdout, r)
+			}
 		}
 
 		if err != nil && err != io.EOF {
@@ -165,11 +239,23 @@ func RunTextCmd(t TaskContext, textStyle dao.Text, prefix string, parallel bool,
 	go func(client Client) {
 		defer wg.Done()
 		var err error
-		if prefix != "" {
-			_, err = io.Copy(os.Stderr, core.NewPrefixer(client.Stderr(), prefix))
+
+		if register == "" {
+			if prefix != "" {
+				_, err = io.Copy(os.Stderr, core.NewPrefixer(client.Stderr(), prefix))
+			} else {
+				_, err = io.Copy(os.Stderr, client.Stderr())
+			}
 		} else {
-			_, err = io.Copy(os.Stderr, client.Stderr())
+			mw := io.MultiWriter(buf, bufErr)
+			r := io.TeeReader(client.Stderr(), mw)
+			if prefix != "" {
+				_, err = io.Copy(os.Stderr, core.NewPrefixer(r, prefix))
+			} else {
+				_, err = io.Copy(os.Stderr, r)
+			}
 		}
+
 		if err != nil && err != io.EOF {
 			fmt.Fprintf(os.Stderr, "%v", err)
 		}
@@ -185,10 +271,10 @@ func RunTextCmd(t TaskContext, textStyle dao.Text, prefix string, parallel bool,
 			fmt.Printf("%s\n", err.Error())
 		}
 
-		return err
+		return buf.String(), bufOut.String(), bufErr.String(), nil
 	}
 
-	return nil
+	return buf.String(), bufOut.String(), bufErr.String(), nil
 }
 
 func HeaderTemplate(header string, data HeaderData) (string, error) {